208 - Added reports functionality
- updated lauch.json file to debug data lake
- updated limit on data lake for each functions
- added since date time for asset lists created by the selected user
- updated inventorydb.go and reportsdb.go file to accomodate query params
- created DetailsCard.jsx component for list of categories and maintenance plans
- updated inventory saga to support params
- updated reports and filter menu to support displaying of selected data. YTD is selected as
default for the reports page when the user lands on the page.

closes #208
