+++ .vscode/launch.json
@@ -10,6 +10,13 @@
             "request": "launch",
             "mode": "debug",
             "program": "${workspaceFolder}/apilayer/"
+        },
+        {
+            "name": "Debug Data Lake Ingest",
+            "type": "go",
+            "request": "launch",
+            "mode": "debug",
+            "program": "${workspaceFolder}/apilayer/dataLake/main.go"
         }
     ]
 }
